Add Close() method to Marathon interface
This implements and required Close for clients (the Marathon
interface), which in turn stops cluster health checks, subscriptions,
and other ongoing behavior.

This lays some groundwork for introducing a context parameter to
methods, but that would need to be a breaking release (adding a new
method to Marathon is breaking enough for anyone aiming to implement
the interface, which is probably nobody, but still worth keeping in
mind).

Clients first stop all operations against themselves before proceeding
to stop all operations on the cluster. Subscriptions are stopped
before marking the client as closed, while the lock is held.